Refactor wtime code into shared routine
